requirements.txt
# Core dependencies
opencv-python>=4.5.0
face-recognition>=1.3.0
numpy>=1.19.0

# GUI dependencies
Pillow>=8.0.0

# Raspberry Pi specific (optional - only needed on Pi)
# Uncomment these lines when installing on Raspberry Pi:
# gpiozero>=1.6.0
# lgpio>=0.2.0

# Note: tkinter is included with Python standard library
